Some French Labels + a change Andrew suggested for CommonEmptyHeader. I commit it since it fix the problem at hand ("&#160;" codes floating all over the place) But this is not as simple as the original issue is due to Labels Manager. If you put a sole space (ie > < ) for CommonEmptyHeader, then if you do some modifications with Labels Manager in this file and then save the file you will get     <property key="CommonEmptyHeader">         <value xml:lang="en"/>     </property>
In order to cope with this I tried to write directly at SaveLabelsToXmlFile.saveLabelsToXmlFile[93] the String "&#160;" which should be ok. But I guess I would have to change the format passed to UtilXml.writeXmlDocument some lines below since else it write "&amp;#160;" and not "&#160;" as intended. Not sure it's possible though. And I have no time to look at it right now.

And we will have to deal with that in a complete way since else we will find an even worst trouble later (as soon someone will use Labels Manager to save changes in this file)



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@767845 13f79535-47bb-0310-9956-ffa450edef68