User friendly message will be displayed to User Interface. More improvements can be done in this commit. Now the message looks like:
In case of success:
The following occurred:
The data exported successfully to eBay.

In case of error:
The Following Errors Occurred:
Input data for tag <Item.Quantity> is invalid or missing. Please check API documentation.

Previously the message was shown as it is in XML form returned from ebay server.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@812213 13f79535-47bb-0310-9956-ffa450edef68