Better French translation, but actually this post is more to speak about the problems we get when we cut sentences and try to use words in translation.
For instance in French, German, Russian, etc.,  all words have a gender (masculine or feminine, not only neutral as in English), actually most languages derived from indo-european language have words gender (http://en.wikipedia.org/wiki/Gender)

As an example : forecast, which is neutral in English, is feminine in French (most words gender are arbitrary, but it's not the point). So when you get something like ${uiLabelMap.CommonCreateNew} ${uiLabelMap.SfaSalesForecast} you are caught in all languages where forecast has a gender :/

Please take this into account because it's much harder afterward to fix.

I will try to post something for contributors in wiki. If you see a related point please chime in

Thanks

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@823903 13f79535-47bb-0310-9956-ffa450edef68