Applied patch from jira issue OFBIZ-3149 - Improve Category association in eBay.
Recently multi store support for Ebay has been implemented. For the multi store support we need ProductStore > ProdCatalog > ProductCategory association. For EBay we define specific categories of type EBAY_CATEGORY, only those products that are associated to an EBay category are eligible for eBay export. In the current implementation there is no proper prodCatalogCategoryTypeId which can be used to associate eBay category to ProdCatalog, one way to associate an eBay category to ProdCatalog is as a PCCT_BROWSE_ROOT category, thus we have two categories associated to the catalog as BROWSE_ROOT, the other one is for online store. Technically, this is wrong, a catalog should be associated with only one browse root category. Due to this wrong association the eBay category doesn't show up on the advance search page of EBAY webapp, instead online store's browse root category is shown which may also contain those products which are not meant to be exported to eBay. So this category associations need to be fixed, and following is how this can be resolved:

   1. Create a new ProdCatalogCatagoryType PCCT_EBAY_ROOT
   2. Define a worker method in CatalogWorker to derive the top level (root) ebay category
   3. Now in the advance search form, first look for top level ebay category associated to the selected catalog
          * If found, derive sub category list of the top level ebay category and display, else display the top level ebay category itself
          * If not found, derive the top level browse root category (PCCT_BROWSE_ROOT) and display the sub category list for it

Thanks Parimal Gain for the contribution.
Special thanks to Mridul Pathak (Zone of Sharp Mind :-)) for discussing implementation plans & providing wonderful notes.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@832713 13f79535-47bb-0310-9956-ffa450edef68