Applied patch from jira issue OFBIZ-3197 - Batch payment issue. On cancelling a payment batch (payment group), payment group members (payments) should available for any another batch. Currently it is not happening.
Proposed solution -
Payment Batches are created by assigning of Payments to FinAccount. In batching FinAccountTrans record is created and then it is assigned to Payment. So on cancelling payment batch if FinAccountTrans status set to cancel. And then association of Payment to FinAccount break and it will available for another batch.

Thanks Sumit for the contribution.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@836129 13f79535-47bb-0310-9956-ffa450edef68