Example: 1. The currency of accounting/facility set to Indian Rupees, its INR 2) Warehouse manager receive a purchase order with an item with price 10 USD 3) In the receive screen, the system prompts for the unit price of 500 INR (the currency exchange rate between INR and USD is about 50 i.e 1 USD -> 50 INR) 4) The user doesn't change the value and after the item is received both the order item (unitPrice) and the supplier product (lastPrice) price are updated with the new unit price of 500 USD.
This is automatically happening with setUnitPriceAsLastPrice service.

With the fix the receive screen now display the original price and its currency, and the price in the facility currency (if different); if the former is edited, and Ajax event update the latter; the latter will be used as inventory item's unit cost; if the former is not updated, then no event will be triggered to change item price or supplier product price.

Bug reported by Jacopo Cappellato and fix provided by Akash Jain. Part of OFBIZ-3389



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@896010 13f79535-47bb-0310-9956-ffa450edef68