Applied patch from OFBIZ-3533 - Associate Image functionality with ProductPromo using content driven. Associate Image functionality with ProductPromo using content driven.
Following work will be covered: -

   1. Add new entity "ProductPromoContent" entity having composite primary key:
         1. productPromoId
         2. contentId
         3. productPromoContentTypeId
         4. fromDate
   2. CRUD services for ProductPromoContent Entity.
   3. Create wrapper ProductPromoContentWrapper
   4. Add content tab at catalog -> promo
   5. List and edit form for associating images to specific productPromoId.
   6. Implement service addImageForProductPromo.

Thanks Amit for the contribution.
Special thanks to Scott for his valuable comments.

PS: Scott regarding to you point # 2 in the last comment, I think it would be fine if we keep association as it in the current patch. Because all the ContentType already exists in the ProductContentType entity. In general we can associate Image, PDF or word document with any promotion so reusing existing entity will be fine IMO. 


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@922506 13f79535-47bb-0310-9956-ffa450edef68