In order to fix issues after issues, I commit this patch from Sascha Rodekamp that fixes "Allow to open a layer lookup from a layer lookup" (https://issues.apache.org/jira/browse/OFBIZ-3446) - OFBIZ-3446
This patch does not fix an issue with calendar in layered lookups. So I will create a new Jira for that.
Also it introduces a small drawback: you can't no longer close a layered lookup with the right left cross button. But Esc. is still working, so it's not a blocker. I will have a look at this and if needed will create a new Jira 







git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@928292 13f79535-47bb-0310-9956-ffa450edef68