A patch from Sascha Rodekamp "Updating Email Address say Contact Mechanism Successfully Updated" (https://issues.apache.org/jira/browse/OFBIZ-3995) - OFBIZ-3995
Sascha: Hm, i had a little bit trouble when i tried to update my address data (shop). The result message was Contact Mechanism Successfully Updated. The service (updateContactMech) tries to differ between email, phone, web ... address changes. But the conditions are always false , because he can't read the from the parameter fields. This patch makes the service read the parameters in the right way.

JLR: I guess the reason we did not spot this before is because OOTB it's only used inside ContactMechServices.xml, and there context is set. Anyway having all in parameters is obviously better/safer, services are not supposed to be used in one situation.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1023842 13f79535-47bb-0310-9956-ffa450edef68