Today in a custom app we experienced a weird issue. We use Ajax (actually "jQuery load") to dynamically load FTL forms which contains lookups. For this to work, we made a small change in the fieldlookup.js script by adding "jQuery live" on the lookup button related click(allow future associations)
It happened that today the lookups build from the load request (Ajax) did not work anymore, nor the other lookups. It was working a week ago and we discovered that was due to the update of jQuery (to 1.4.4 this week and 1.4.3 last week). So I have reverted the calls to the jQuery lib to 1.4.2. I will report the problem to the jQuery team (nothing like that seems to be known for now on the Internet).

This commit introduces 
* the "jQuery live" change in fieldlookup.js script 
* revert calls to the jQuery lib from 1.4.4 version to 1.4.2
* in order to help investigations put back jQuery libs 1.4.2 and 1.4.3

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/jquery@1036154 13f79535-47bb-0310-9956-ffa450edef68