A patch from Anne Jessel "Is Empty for text-find widget broken" (https://issues.apache.org/jira/browse/OFBIZ-4025) - OFBIZ-4025
To reproduce:
    * go to Facility, Inventory Items tab (or any form that uses text-find widget)
    * choose "Is Empty" from drop-down next to Product Id
    * click Find button

Expected behaviour: no matching items listed, as all items have a Product Id.
Observed behaviour: all items match.

Problem is caused by find code assuming that there is no entity condition to be created if the text field is empty, but of course Is Empty is
different to the other operators in that the text field is expected to be empty.

The attached patch adds a special check for the Is Empty operator.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1042396 13f79535-47bb-0310-9956-ffa450edef68