Fix "Ajax requests prevent externalLoginKey parameters from working correctly" (https://issues.apache.org/jira/browse/OFBIZ-3862) - OFBIZ-3862
Scott reported: To clarify, the problem only occurs if the Ajax call results in a screen being rendered since it is the ScreenRenderer that causes a new key to be generated.

Bilgin suggested: I think a solution would be to skip generating of new external login key for Ajax requests.
Ajax requests can be identified by presence of 'X-Requested-With': 'XMLHttpRequest', http header. It is set by most of the Javascript libraries we use: prototype, jquery.

I coded it

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1068283 13f79535-47bb-0310-9956-ffa450edef68