A patch from Stephen Rufle "Checkout Shipping Options" (https://issues.apache.org/jira/browse/OFBIZ-4162) - OFBIZ-4162
When the user selects a shipping option and moves to the next step and then comes back to the "Shipping Options" step the previously selected value is not selected on the screen. This is because in the ftl a comparison is made between shippingMethod and chosenShippingMethod. chosenShippingMethod is being escaped so the values will never match. The solution is to use wrapString on chosenShippingMethod.

# Add a product to the cart
# login as any user I used "admin"
# Checkout Step "Shipping Address" (Step 1: Where shall we ship it?)
    * Click Next
# Checkout Step "Shipping Options" (Step 2: How shall we ship it?)
    * Choose "UPS Air"
    * Click Next
# Checkout Step "Payment Options" (Step 3: How shall you pay?)
    * Click "Browser" back button, in my code had a request that directed back a step, but this shows the issue
# ISSUE: Now "UPS Air" is no longer selected, after my patch it will be

JLR: this was only with modified code, but it makes sense and can't hurt. I will not backport though

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1068286 13f79535-47bb-0310-9956-ffa450edef68