A modified patch from Pierre TALLOTTE "Impossible to add a variant product in a purchase order" (https://issues.apache.org/jira/browse/OFBIZ-4166) - OFBIZ-4166
When we add a variant product in a purchase order, we get an error. This error occurs in ProductDetail.groovy because virtualPriceMap.basePrice is null in Purchase order case.

JLR: Pierre did the analysis and proposed a patch, I reviewed and finally use a patch of my own since Pierre's was not handling both cases (sales AND purchase)


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1070982 13f79535-47bb-0310-9956-ffa450edef68