A modified patch from Konstantin Nikkel "Exception while creating a productBarCode" https://issues.apache.org/jira/browse/OFBIZ-4226 - OFBIZ-4226
If you try to create a barcode of a product that has no productName as content, you get the exception "Variable pruductName does not exist".
An exclamation mark in the "ProductBarCode.fo.ftl" manages this problem.

I added this feature, so I changed it a bit more by using also internalName as default for productName and changed the decorator since now productName is not needed (better to do that only in the ProductBarCode.fo screen, I checked it's no used anywhere else)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1084671 13f79535-47bb-0310-9956-ffa450edef68