Closes <<Caching is used by org.ofbiz.webapp.control.ProtectViewWorker(line 68) for UserLoginAndProtectedView (since changeset r1075350), but the view has set never-cache="true">> reported by Jonas Hoef at https://issues.apache.org/jira/browse/OFBIZ-4243
I wondered if I should not remove the event from the common-controller since I guess this is rarely used. But, as the functionality is available OOTB, I preferred to keep it as is. So I simply remove the never-cache from  UserLoginAndProtectedView view-entity. Since it will almost ever be empty, no memory issues and better performance is expected. If I remember well, at the moment I committed this, I checked: the performance impact was already not noticeable (only one check).

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1100197 13f79535-47bb-0310-9956-ffa450edef68