after session timeout, ajax popup dialogbox shows ofbiz login screen (https://issues.apache.org/jira/browse/OFBIZ-4256).
If the session is timed out the request handler checks weather it is an ajax request or not (The information comes from the request header if the X-Requested-With contains XMLHttpRequest we haven an ajax request.).
If it is an ajax request it don't return the normal "checkLogin" request, i return a a special ajaxCheckLogin request. This request can point to an site which shows the "Please Login Hint".

Orig Msg:
Take the following page (for example) https://localhost/content/control/findContent

field "Data Resource Id" has an icon beside it that creates an ajax dialog pop up. If the session has already timed out. The dialog box would show the login screen.

I suspect that all pages containing this functionality would exhibit the same symptom.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1102615 13f79535-47bb-0310-9956-ffa450edef68