I got caught by this typo on a production cluster: XmlFileLoader: File file:.....framework/entity/config/entityengine.xml process fatal error. Line: 686. Error message: Invalid byte 2 of 2-byte UTF-8 sequence.
After having put back the line modified by a patch automatically applied, I got this  mis-representation. Not sure why, and anyway had no chances to think about it (certainly due to a difference on how UTF-8 was represented on these servers and on my machine - XP with Putty and SSH)

So I remove this weirdness to prevent future possible nervous crises.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1104173 13f79535-47bb-0310-9956-ffa450edef68