Refactoring: Now useing the jackrabbit ObjectContentMapping Library to map POJOs to the Jackrabbit repository. That makes the code more generic and we can define Content Objects for different approches (ie. File Content, News Content, Blog content, Forum Content, Product Content ...).
Furthermore i removed the connection between the repository and the database. The default JCR implementation went back to the framework layer and can be used from each application. If an application need additional content information in the database, it is possible to build a database access layer on top of the default JCR implementation (application specific).


TODO: Versioning, Language Support, Better Tests, documentation

  

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/jackrabbit20100709@1156196 13f79535-47bb-0310-9956-ffa450edef68