A patch from Paul Foxworthy "roleTypeId not copied from ProductStoreShipmentMethod to OrderItemShipGroup" https://issues.apache.org/jira/browse/OFBIZ-4389
In ordershippinginfo.ftl, a list of shipping methods is presented, and when one is chosen values are written into an OrderItemShipGroup. The shipping methods are defined by rows in ProductStoreShipMethod, and one of the attributes of this entity is a carrierRoleTypeId. Similarly, OrderItemShipGroup has a carrierRoleTypeId, but no data is being written into it.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1167116 13f79535-47bb-0310-9956-ffa450edef68