A patch from Paul Foxworthy https://issues.apache.org/jira/browse/OFBIZ-4784 "storeIncomingEmail assumes variable is not null when it might be"
In storeIncomingEmail in CommunicationEventServices.java, the code to assemble the aboutThisEmail debugging message assumes addressesFrom and addressesTo are not null. But the methods that return values for those variables might return null if they encounter an exception.

Removes also uses of UtilValidate.isEmpty, which should only be used for collections and sequences (see OFBIZ-4427).

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1309715 13f79535-47bb-0310-9956-ffa450edef68